refactor(icon-button): to TypeScript, use readme generator
