+++ packages/components/buttons/flat-button/README.md
@@ -51,7 +51,7 @@ export default Example;
 
 | Props          | Type                                                                    | Required | Default     | Description                                                                                                                                                |
 | -------------- | ----------------------------------------------------------------------- | :------: | ----------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `as`           | `union`<br/>Possible values:<br/>`string , ComponentType`               |          |             | You may pass in a string like "a" to have the button render as an anchor tag instead.&#xA;<br/>&#xA;Or you could pass in a React Component, like a `Link`. |
+| `as`           | `union`<br/>Possible values:<br/>`string , ElementType`                 |          |             | You may pass in a string like "a" to have the button render as an anchor tag instead.&#xA;<br/>&#xA;Or you could pass in a React Component, like a `Link`. |
 | `tone`         | `union`<br/>Possible values:<br/>`'primary' , 'secondary' , 'inverted'` |          | `'primary'` | Indicates the color scheme of button                                                                                                                       |
 | `type`         | `union`<br/>Possible values:<br/>`'submit' , 'reset' , 'button'`        |          | `'button'`  | Used as the HTML `type` attribute.                                                                                                                         |
 | `label`        | `string`                                                                |    ✅    |             | Should describe what the button is for                                                                                                                     |
