+++ packages/components/buttons/icon-button/docs/additional-info.md
@@ -1,0 +1,4 @@
+### Examples in the Merchant Center
+
+- Secondary action example: Delete product
+- Minimize effect example: Reordering table
