+++ packages/components/buttons/icon-button/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@commercetools-uikit/icon-button",
-  "description": "Icon buttons are icon-only buttons.",
+  "description": "Icon buttons are icon-only buttons. They trigger an action when clicked (`onClick` prop). You must also pass a label for accessibility reasons.",
   "version": "12.0.0",
   "bugs": "https://github.com/commercetools/ui-kit/issues",
   "repository": {
